requirements.txt
prometheus-client
pyyaml
confuse
requests
openstacksdk
pyjwt
PyYAML
